+++ packages/fetcher/src/factories.ts
@@ -1,4 +1,3 @@
-import { GetServerSidePropsContext } from 'next'
 import { generateUrl, parseUrl } from '@titicaca/view-utilities'
 
 import { HttpResponse, RequestOptions } from './types'
@@ -169,28 +168,4 @@ export function authFetcherize<Fetcher extends BaseFetcher>(
 
     return secondTrialResponse
   }) as unknown as ExtendFetcher<Fetcher, typeof NEED_LOGIN_IDENTIFIER>
-}
-
-export function i18nFetcherize<Fetcher extends BaseFetcher>(
-  fetcher: Fetcher,
-  ctx: GetServerSidePropsContext,
-): Fetcher {
-  const {
-    req: { headers },
-  } = ctx
-  const langHeader = (headers['x-triple-user-lang'] ?? 'ko') as string
-  const countryHeader = (headers['x-triple-user-country'] ?? 'kr') as string
-
-  return ((href, optionsParams) => {
-    const { headers, ...options } = optionsParams ?? {}
-
-    return fetcher(href, {
-      ...options,
-      headers: {
-        ...headers,
-        'x-triple-user-lang': langHeader,
-        'x-triple-user-country': countryHeader,
-      },
-    })
-  }) as Fetcher
 }
